bot_army v0.1.0 BotArmy.Actions

Generic Actions.

Actions are functions that take the bot's context and any supplied arguments, perform some useful side effects, and then return the outcome. The context is always passed as the first argument.

Valid outcomes are: :succeed, :fail, :continue, :done or {:error, reason}.

:succeed, :fail, and :continue can also be in the form of {:succeed, key: "value"} if you want save/update the context.

Types

outcome()

outcome() ::
  :succeed
  | :fail
  | :continue
  | :done
  | {:error, any()}
  | {:succeed, keyword()}
  | {:fail, keyword()}
  | {:continue, keyword()}

Actions must return one of these outcomes.

Functions

action(module, fun, args \\ [])

A semantic helper to define actions in your behavior tree.

Node.sequence([
  ...
  action(BotArmy.Actions, :wait, [5]),
  ...
  action(BotArmy.Actions, :done)
])

done(_)

This will stop the bot from running (by default bots "loop" continously through their behavior trees

error(_, reason)

Signal that this bot has errored, causing the bot's process to die with the given reason.

log(context, message)

A helpful way to "tap" the flow of the behavior tree for debugging.

succeed_rate(context, rate)

Given a rate as a percentage, this will succeed that percent of the time, and fail otherwise.

For example succeed_rate(context, 0.25) will succeed on average 1 our of 4 tries.

wait(context, s \\ 5)

Makes the calling process wait for the given number of seconds

wait(context, min, max)

Makes the calling process wait for a random number of seconds in the range defined by the given integers min and max