bot_army v0.1.0 BotArmy.IntegrationTest.FSMState protocol

Types

t()

t() :: term()

transition_response()

transition_response() ::
  {:next_fsm_state, fsm_state :: BotArmy.IntegrationTest.FSMState.t(),
   response :: term(), BotArmy.IntegrationTest.t()}
  | {:keep_fsm_state, response :: term(), BotArmy.IntegrationTest.t()}

Response from a state specifying which state to transition to (or to keep the current state), and the response to send to the caller, and the new state.

Functions

on_enter(fsm_state, state)

work to do upon entering a state

run(fsm_state, opts, state)

start a new test run

test_failed(fsm_state, test_ref, reason, state)

test failed

test_succeeded(fsm_state, test_ref, state)

test passed