bot_army v0.1.0 BotArmy.IntegrationTest.Workflow behaviour

A behaviour to implement when defining an integration test run.

This behaviour allows defining multiple trees to run in parallel, as well as a "pre" and "post" tree.

For example, a "pre" tree might check the server health and obtain a log in token to store in BotArmy.SharedData. Then multiple trees could use the same token to run various tests in parallel. Finally, a "post" tree could do a final check or cleanup.

Note that each parallel tree will be run by a new bot instance. Keep race conditions in mind if your tests make use of the same resource. If any tree fails (including the pre and post trees), the entire run will fail. The post tree will always run, even if prior trees fail

You can define each stage directly by implementing the callbacks, or you can use the supplied macros.

defmodule MyWorkflow do
  use Workflow

  pre Node.always_succeed(action(Actions, :log, ["pre"]))

  post Node.always_succeed(action(Actions, :log, ["post"]))

  parallel "test 1", Node.always_succeed(action(Actions, :log, ["first parallel"]))

  parallel "test 2", Node.always_succeed(action(Actions, :wait, [0]))

  # do syntax variant
  parallel "test 3" do
    var = "runtime"

    Node.sequence([
      action(Actions, :log, [var]),
      action(Actions, :wait, [0])
    ])
  end
end

Functions

parallel(name, contents) (macro)

A simple way to define each test you want to run in parallel. Add as many of these as you want in your workflow module.

You can specify a name and node, or use do syntax if you need to run other code to set up the tree.

parallel "test xyz", xyz_test_node

parallel "test abc" do
  n = :rand.uniform()
  abc_test_node(n)
end

You must have at least one of these, and all test names must be unique.

post(node) (macro)

A simple way to define your "post" test.

post clean_up_tree

This is optional, and may only be defined once.

pre(node) (macro)

A simple way to define your "pre" test.

pre setup_tree

This is optional, and may only be defined once.

Callbacks

parallel()

parallel() :: %{required(any()) => BehaviorTree.Node.t()}

Required. A map of trees which will each be run in parallel with their own bot.
The key is used for reporting.

post()

post() :: BehaviorTree.Node.t()

Optional. A tree to run after all other trees in this run have completed, or if any tree fails.

This is useful for tear down or testing postconditions.

pre()

pre() :: BehaviorTree.Node.t()

Optional. A tree to run before doing anything else in this run.

This is useful for set up or testing preconditions.